requirements.txt
dicttoxml==1.7.4
Django==1.11
django-admin-json-editor==0.1.5
django-model-utils==3.1.1
django-nose==1.4.5
jsonschema==2.6.0
nose==1.3.7
psycopg2==2.7.4
requests==2.18.4
